pass ImageBitmap and imageP5 if p5 is present
Now pass an imageBitmap for fragments in the metacollector object. ImageBitmap is a native JavaScript interface to store and draw images on canvas. It is polyfilled, in particular for on Safari <= 14

The goal is to make it easy for generative artists who don't use p5 to draw the image on the canvas, using the native CanvasRenderingContext2D.drawImage() function

The image property that was holding a p5 image object is now replaced by an imageP5 property, that is only added if p5 is present in the global context. That way when using p5.js, generative artists just have to use this property and get all the usual features.